Add PositionResetter to Paddles
It's possible for paddles to move horizontally in
response to collisions with the puck.  Using
MoveAndCollide instead of MoveAndSlide makes this less
likely, but it can still happen.  It became necessary
to inject a component to the base Paddle that would
reset its X position, so that the game would not
become unplayable in the event that the PlayerPaddle
wandered off screen.